+++ CHANGELOG.md
@@ -32,6 +32,8 @@
 - Fix `aligned_free` when used with `FLATCC_USE_GENERIC_ALIGNED_ALLOC` (#118)
 - Fix potential buffer overrun when parsing JSON containing surrogate pairs
   with a resulting UTF-8 code point of length 4 (bug introduced in 0.5.3).
+- BREAKING: empty structs are no longer supported. They were fully supported
+  and Googles flatc compiler also no longer support them.
 
 ## [0.5.3]
 - BREAKING: 0.5.3 changes behavour of builder create calls so arguments